requirements.txt
matplotlib==3.5.0
pandas==1.3.4
python-dotenv==0.19.2
requests==2.26.0
seaborn==0.11.2
